requirements.txt
pandas
hydra-core
numpy
omegaconf
scikit-learn
opencv-python
tqdm
gdown